refactor: set the FOLDER STRUCTURE
