--- src/core/api/index.ts
@@ -1,1 +1,0 @@
-// 마이그레이션 이전 파일 중에 utils/lib/index.ts 옮겨오면 될듯
