+++ src/core/atom.ts
@@ -1,8 +1,7 @@
 import { atom } from "recoil";
 import { recoilPersist } from "recoil-persist";
 
-import { NavigatingBookInfoState } from "../dataType";
-
+import { NavigatingBookInfoState } from "../types/bookcase";
 const { persistAtom } = recoilPersist();
 
 export const isLoginState = atom<boolean>({
