+++ src/util/exampleData.ts
@@ -1,0 +1,8 @@
+/*
+마지막 편집자: 22-04-28 joohaem
+변경사항 및 참고:
+  - core 폴더로 옮기는 것 고려해주십시오!
+    
+고민점:
+  - 
+*/
