+++ tsconfig.json
@@ -18,6 +18,6 @@
     "noFallthroughCasesInSwitch": true,
     "noImplicitAny": true,
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "src/util/hooks/useCheckLoginState.tsㅌ"],
   "exclude": ["node_modules"]
 }
